requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.5
aiohttp-socks==0.8.4
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
cssselect==1.2.0
frozenlist==1.4.1
idna==3.7
jmespath==1.0.1
lxml==5.2.2
magic-filter==1.0.12
multidict==6.0.5
packaging==24.1
parsel==1.9.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
python-socks==2.5.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
w3lib==2.2.1
yarl==1.9.4